+++ scripts/config/constants.config.mjs
@@ -2,8 +2,8 @@ export const MODULE_ID = "simplecover5e";
 
 export const COVER_STATUS_IDS = {
   half: "coverHalf",
-  threeQuarters: "coverThreeQuarters"
-  //total: "coverTotal"
+  threeQuarters: "coverThreeQuarters",
+  total: "coverTotal"
 };
 
 export const SETTING_KEYS = {